# openapi.yaml
openapi: 3.1.0
info:
  title: User Service API
  version: 1.0.0
  description: OpenAPI спецификация для основного сервиса
security:
  - bearerAuth: [ ]
paths:
  /api/v1/main_service/projects/{projectId}/subscriptions:
    post:
      tags:
        - ProjectSubscription
      summary: Подписка на проект
      description: Подписывает текущего пользователя на указанный проект.
      operationId: subscribeToProject
      parameters:
        - name: projectId
          in: path
          required: true
          description: Идентификатор проекта
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Подписка успешно выполнена
          content: { }
    delete:
      tags:
        - ProjectSubscription
      summary: Отписка от проекта
      description: Отписывает текущего пользователя от указанного проекта.
      operationId: unsubscribeFromProject
      parameters:
        - name: projectId
          in: path
          required: true
          description: Идентификатор проекта
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Пользователь успешно отписан
          content: { }
    get:
      tags:
        - ProjectSubscription
      summary: Проверка подписки пользователя
      description: Проверяет, подписан ли текущий пользователь на проект.
      operationId: isUserSubscribed
      parameters:
        - name: projectId
          in: path
          required: true
          description: Идентификатор проекта
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Статус подписки
          content:
            application/json:
              schema:
                type: boolean
                example: true
  /api/v1/main_service/projects/{projectId}/surveys:
    post:
      tags:
        - ProjectSurvey
      summary: Создание опроса для проекта
      description: Создает новый опрос для указанного проекта.
      operationId: createProjectSurvey
      parameters:
        - name: projectId
          in: path
          required: true
          description: Идентификатор проекта
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectSurveyRequest'
      responses:
        '201':
          description: Опрос успешно создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectSurveyFullResponse'
    get:
      tags:
        - ProjectSurvey
      summary: Получение опроса проекта
      description: Возвращает существующий опрос для указанного проекта.
      operationId: getProjectSurvey
      parameters:
        - name: projectId
          in: path
          required: true
          description: Идентификатор проекта
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Информация об опросе
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectSurveyFullResponse'
    delete:
      tags:
        - ProjectSurvey
      summary: Удаление опроса проекта
      description: Удаляет опрос для указанного проекта.
      operationId: deleteProjectSurvey
      parameters:
        - name: projectId
          in: path
          required: true
          description: Идентификатор проекта
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Опрос успешно удален
          content: { }
  /api/v1/main_service/projects/{projectId}/surveys/submissions:
    post:
      tags:
        - ProjectSurvey
      summary: Отправка ответов на опрос
      description: Отправляет ответы текущего пользователя на опрос проекта.
      operationId: submitProjectSurveyAnswers
      parameters:
        - name: projectId
          in: path
          required: true
          description: Идентификатор проекта
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SurveySubmissionRequest'
      responses:
        '201':
          description: Ответы успешно отправлены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SurveySubmissionFullResponse'
    get:
      tags:
        - ProjectSurvey
      summary: Получение всех отправленных ответов
      description: Возвращает список всех отправленных ответов на опрос проекта. Поддерживается сортировка по параметру `sort`.
      operationId: getProjectSurveySubmissions
      parameters:
        - name: projectId
          in: path
          required: true
          description: Идентификатор проекта
          schema:
            type: integer
            format: int64
        - name: sort
          in: query
          required: false
          description: Параметр сортировки (например, по дате отправки)
          schema:
            type: string
      responses:
        '200':
          description: Список ответов на опрос
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SurveySubmissionFullResponse'
...
  /api/v1/main_service/projects:
    post:
      tags:
        - Project
      summary: Создание проекта
      description: Создает новый проект. Использует имя пользователя из контекста безопасности.
      operationId: createProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectRequest'
      responses:
        '201':
          description: Проект успешно создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectFullResponse'
    get:
      tags:
        - Project
      summary: Получение списка проектов
      description: Возвращает список проектов с поддержкой фильтрации по категории, поисковому запросу и статусу модерации. Поддерживает постраничный вывод.
      operationId: getProjects
      parameters:
        - name: category
          in: query
          description: Фильтр по категории проекта
          schema:
            type: string
        - name: query
          in: query
          description: Поисковый запрос по названию проекта
          schema:
            type: string
        - name: moderationPassed
          in: query
          description: Фильтр по статусу модерации
          schema:
            type: boolean
        - name: page
          in: query
          description: Номер страницы
          required: true
          schema:
            type: integer
            example: 0
        - name: size
          in: query
          description: Размер страницы
          required: true
          schema:
            type: integer
            example: 10
      responses:
        '200':
          description: Список проектов
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectPageResponse'
  /api/v1/main_service/projects/{id}:
    get:
      tags:
        - Project
      summary: Получение проекта по ID
      description: Возвращает полную информацию о проекте по идентификатору.
      operationId: getProject
      parameters:
        - name: id
          in: path
          required: true
          description: Идентификатор проекта
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Информация о проекте
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectFullResponse'
    patch:
      tags:
        - Project
      summary: Обновление проекта
      description: Обновляет данные проекта по идентификатору.
      operationId: updateProject
      parameters:
        - name: id
          in: path
          required: true
          description: Идентификатор проекта
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectRequest'
      responses:
        '200':
          description: Информация о проекте
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectFullResponse'
    delete:
      tags:
        - Project
      summary: Удаление проекта
      description: Удаляет проект по идентификатору.
      operationId: deleteProject
      parameters:
        - name: id
          in: path
          required: true
          description: Идентификатор проекта
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Проект успешно удален
  /api/v1/main_service/projects/{id}/likes:
    post:
      tags:
        - Project
      summary: Лайк/дизлайк проекта
      description: Переключает лайк текущего пользователя для указанного проекта.
      operationId: toggleProjectLike
      parameters:
        - name: id
          in: path
          required: true
          description: Идентификатор проекта
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Состояние лайка после операции
          content:
            application/json:
              schema:
                type: boolean
                example: true
  /api/v1/main_service/projects/{id}/likes/count:
    get:
      tags:
        - Project
      summary: Получение количества лайков проекта
      description: Возвращает текущее количество лайков проекта.
      operationId: getProjectLikesCount
      parameters:
        - name: id
          in: path
          required: true
          description: Идентификатор проекта
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Количество лайков проекта
          content:
            application/json:
              schema:
                type: integer
                format: int64
  /api/v1/main_service/projects/{id}/moderationPassed:
    patch:
      tags:
        - Project
      summary: Обновление статуса модерации проекта
      description: Устанавливает статус прохождения модерации для проекта.
      operationId: updateProjectModerationStatus
      parameters:
        - name: id
          in: path
          required: true
          description: Идентификатор проекта
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: boolean
              example: true
      responses:
        '200':
          description: Статус модерации успешно обновлен
...
components:
  schemas:
...
    ProjectPageResponse:
      allOf:
        - $ref: '#/components/schemas/Page'
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/ProjectSummaryResponse'
...
    Page:
      type: object
      required:
        - content
        - number
        - size
        - totalElements
        - totalPages
      properties:
        content:
          type: array
          description: Массив элементов текущей страницы
          items: { }
        number:
          type: integer
          example: 1
        size:
          type: integer
          example: 1
        totalElements:
          type: integer
          format: int64
          example: 1
        totalPages:
          type: integer
          example: 1
    UserFullResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Уникальный идентификатор пользователя
          example: 42
        username:
          type: string
          description: Имя пользователя
          example: "Andrey"
    ProjectMemberFullResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Идентификатор участника проекта
          example: 15
        username:
          type: string
          description: Имя пользователя
          example: "User12345"
        role:
          type: string
          description: Роль пользователя в проекте
          example: "DevOps"
    CommentFullResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Идентификатор комментария
          example: 1001
        owner:
          $ref: '#/components/schemas/UserFullResponse'
        text:
          type: string
          description: Текст комментария
          example: "User12345"
        createdDate:
          type: string
          format: date-time
          description: Дата и время создания комментария (UTC)
          example: "2026-01-24T14:32:10"
    ProjectMemberRequest:
      type: object
      required:
        - username
        - role
      properties:
        username:
          type: string
          description: Имя пользователя
          minLength: 1
          example: "User12345"
        role:
          type: string
          description: Роль пользователя в проекте
          minLength: 1
          example: "DevOps"
    ProjectFullResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Идентификатор проекта
          example: 101
        projectOwner:
          $ref: '#/components/schemas/UserFullResponse'
        members:
          type: array
          description: Участники проекта
          items:
            $ref: '#/components/schemas/ProjectMemberFullResponse'
        projectName:
          type: string
          description: Название проекта
          example: "Open Source Platform"
        projectDescription:
          type: string
          description: Описание проекта
          example: "Платформа для совместной разработки open-source проектов"
        projectLogoUrl:
          type: string
          format: uri
          description: URL логотипа проекта
          example: "https://cdn.example.com/logos/project.png"
        category:
          type: string
          description: Категория проекта
          example: "Technology"
        likes:
          type: integer
          format: int64
          description: Количество лайков проекта
          example: 245
        hasSurvey:
          type: boolean
          description: Наличие опроса в проекте
          example: true
        comments:
          type: array
          description: Комментарии к проекту
          items:
            $ref: '#/components/schemas/CommentFullResponse'
    ProjectSummaryResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Идентификатор проекта
          example: 101
        projectOwner:
          type: string
          description: Имя пользователя
          example: "User12345"
        projectName:
          type: string
          description: Название проекта
          example: "Open Source Platform"
        projectLogoUrl:
          type: string
          format: uri
          description: URL логотипа проекта
          example: "https://cdn.example.com/logos/project.png"
        category:
          type: string
          description: Категория проекта
          example: "Technology"
        likes:
          type: integer
          format: int64
          description: Количество лайков проекта
          example: 245
    ProjectRequest:
      type: object
      required:
        - projectName
        - projectDescription
        - projectCategoryId
      properties:
        members:
          type: array
          description: Участники проекта
          items:
            $ref: '#/components/schemas/ProjectMemberRequest'
        projectName:
          type: string
          description: Название проекта
          minLength: 1
          example: "Open Source Platform"
        projectDescription:
          type: string
          description: Описание проекта
          minLength: 1
          example: "Платформа для совместной разработки open-source проектов"
        projectLogoUrl:
          type: string
          format: uri
          description: URL логотипа проекта
          example: "https://cdn.example.com/logos/project.png"
        projectCategoryId:
          type: integer
          format: int64
          description: Идентификатор категории проекта
          example: 3
    ProjectSurveyRequest:
      type: object
      required:
        - questions
      properties:
        questions:
          type: array
          description: Список вопросов опроса проекта
          minItems: 1
          items:
            $ref: '#/components/schemas/ProjectSurveyQuestionRequest'
    ProjectSurveyQuestionRequest:
      type: object
      required:
        - questionText
      properties:
        questionText:
          type: string
          description: Текст вопроса
          minLength: 1
          example: "Какую функциональность вы хотели бы видеть в следующей версии?"
    SurveySubmissionRequest:
      type: object
      required:
        - answers
      properties:
        answers:
          type: array
          description: Список ответов пользователя на вопросы опроса
          minItems: 1
          items:
            $ref: '#/components/schemas/SurveyAnswerRequest'
    SurveyAnswerRequest:
      type: object
      required:
        - questionId
        - answerText
      properties:
        questionId:
          type: integer
          format: int64
          description: Идентификатор вопроса опроса
          example: 12
        answerText:
          type: string
          description: Текст ответа пользователя
          minLength: 1
          example: "Хотелось бы видеть мобильное приложение"
    ProjectSurveyFullResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Идентификатор опроса
          example: 25
        projectId:
          type: integer
          format: int64
          description: Идентификатор проекта
          example: 101
        questions:
          type: array
          description: Список вопросов опроса
          items:
            $ref: '#/components/schemas/SurveyQuestionFullResponse'
    SurveySubmissionFullResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Идентификатор отправки ответов
          example: 55
        surveyId:
          type: integer
          format: int64
          description: Идентификатор опроса
          example: 10
        respondentUsername:
          type: string
          description: Username пользователя, прошедшего опрос
          example: "User12345"
        answers:
          type: array
          description: Ответы пользователя на вопросы опроса
          items:
            $ref: '#/components/schemas/SurveyAnswerFullResponse'
        averageRating:
          type: integer
          description: Средняя оценка по опросу (если применимо)
          example: 4
    SurveyQuestionFullResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Идентификатор вопроса
          example: 1
        questionText:
          type: string
          description: Текст вопроса
          example: "Насколько вам интересен этот проект?"
    SurveyAnswerFullResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Идентификатор ответа
          example: 25
        questionText:
          type: string
          description: Текст вопроса, на который дан ответ
          example: "Насколько вам интересен этот проект?"
        answerText:
          type: string
          description: Текст ответа пользователя
          example: "Очень интересен"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT