# src/main/java/org/siri_hate/main_service/controller/ProjectController.java
package org.siri_hate.main_service.controller;

import org.siri_hate.main_service.api.ProjectApi;
import org.siri_hate.main_service.dto.ProjectFullResponseDTO;
import org.siri_hate.main_service.dto.ProjectPageResponseDTO;
import org.siri_hate.main_service.dto.ProjectRequestDTO;
import org.siri_hate.main_service.service.ProjectService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.security.core.Authentication;
import org.springframework.security.core.context.SecurityContextHolder;
import org.springframework.web.bind.annotation.*;

@RestController
public class ProjectController implements ProjectApi {

    private final ProjectService projectService;

    @Autowired
    public ProjectController(ProjectService projectService) {
        this.projectService = projectService;
    }

    @Override
    public ResponseEntity<ProjectFullResponseDTO> createProject(ProjectRequestDTO projectRequestDTO) {
        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
        String ownerUsername = authentication.getName();
        var response = projectService.createProject(ownerUsername, projectRequestDTO);
        return new ResponseEntity<>(response, HttpStatus.OK);
    }

    @Override
    public ResponseEntity<Void> deleteProject(Long id) {
        projectService.deleteProject(id);
        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
    }

    @Override
    public ResponseEntity<ProjectFullResponseDTO> getProject(Long id) {
        var response = projectService.getProject(id);
        return new ResponseEntity<>(response, HttpStatus.OK);
    }

    @Override
    public ResponseEntity<Long> getProjectLikesCount(Long id) {
        var response = projectService.getProjectLikesCount(id);
        return new ResponseEntity<>(response, HttpStatus.OK);
    }

    @Override
    public ResponseEntity<ProjectPageResponseDTO> getProjects(Integer page, Integer size, String category, String query, Boolean moderationPassed) {
        var response = projectService.getProjects(category, query, page, size, moderationPassed);
        return new ResponseEntity<>(response, HttpStatus.OK);
    }

    @Override
    public ResponseEntity<Boolean> toggleProjectLike(Long id) {
        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
        String username = authentication.getName();
        projectService.toggleProjectLike(username, id);
        return new ResponseEntity<>(HttpStatus.OK);
    }

    @Override
    public ResponseEntity<ProjectFullResponseDTO> updateProject(Long id, ProjectRequestDTO projectRequestDTO) {
        var response = projectService.updateProject(id, projectRequestDTO);
        return new ResponseEntity<>(response, HttpStatus.OK);
    }

    @Override
    public ResponseEntity<Void> updateProjectModerationStatus(Long id, Boolean body) {
        projectService.updateProjectModerationStatus(id, body);
        return new ResponseEntity<>(HttpStatus.OK);
    }
}